refactor: Counter에 suspense 적용
- 서버 변경 시 두 번 깜빡이는 문제 개선
